dash-app/requirements.txt
numpy<2
dash
dash_bootstrap_components
plotly
yfinance
cufflinks
requests
gunicorn
python-dotenv
statsmodels
